fix: don't trigger email notification send
