+++ src/routers/auth/signup.ts
@@ -1,7 +1,7 @@
 import { NextFunction, Request, Response } from "express";
 import prisma from "@mirlo/prisma";
 import { hashPassword } from ".";
-import sendMail from "../../queues/send-mail";
+import sendMail from "../../jobs/send-mail";
 import { NotificationType } from "@mirlo/prisma/client";
 import { Job } from "bullmq";
